Move staking components into staking module
