+++ packages/app/src/common/components/index.ts
@@ -7,9 +7,5 @@ export * from "./LockedGSol";
 export * from "./LockForm";
 export * from "./Panel";
 export * from "./Spinner";
-export * from "./StakeForm";
 export * from "./TooltipPopover";
 export * from "./TweetButton";
-export * from "./UnstakeForm";
-export * from "./UnstakeOption";
-export * from "./WithdrawTickets";
