+++ packages/app/src/staking/pages/StakeDashboard.tsx
@@ -17,13 +17,10 @@ import {
   DetailsBox,
   Panel,
   Button,
-  UnstakeForm,
   InfoBox,
-  WithdrawTicket,
   TooltipPopover,
   LockForm,
   LockedGSol,
-  StakeForm,
 } from "../../common/components";
 import { useSunriseStake } from "../../common/context/sunriseStakeContext";
 import {
@@ -34,6 +31,7 @@ import {
 import { useCarbon } from "../../common/hooks/useCarbon";
 import { tooltips } from "../../common/content/tooltips";
 import { type SunriseClientWrapper } from "../../common/sunriseClientWrapper";
+import { StakeForm, UnstakeForm, WithdrawTicket } from "../components";
 
 const StakeDashboard: FC = () => {
   const wallet = useWallet();
